IGNITE-26787 Thread Pools related classes moved to internal package
