+++ modules/core/src/main/java/org/apache/ignite/internal/processors/pool/PoolProcessor.java
@@ -45,10 +45,13 @@ import org.apache.ignite.internal.processors.plugin.IgnitePluginProcessor;
 import org.apache.ignite.internal.processors.security.IgniteSecurity;
 import org.apache.ignite.internal.processors.timeout.GridTimeoutProcessor;
 import org.apache.ignite.internal.systemview.StripedExecutorTaskViewWalker;
+import org.apache.ignite.internal.thread.pool.IgniteStripedThreadPoolExecutor;
+import org.apache.ignite.internal.thread.pool.IgniteThreadPoolExecutor;
 import org.apache.ignite.internal.thread.pool.OperationContextAwareIoPool;
 import org.apache.ignite.internal.thread.pool.OperationContextAwareStripedExecutor;
 import org.apache.ignite.internal.thread.pool.OperationContextAwareStripedThreadPoolExecutor;
 import org.apache.ignite.internal.thread.pool.OperationContextAwareThreadPoolExecutor;
+import org.apache.ignite.internal.thread.pool.SameThreadExecutor;
 import org.apache.ignite.internal.util.StripedExecutor;
 import org.apache.ignite.internal.util.tostring.GridToStringExclude;
 import org.apache.ignite.internal.util.typedef.F;
@@ -61,9 +64,6 @@ import org.apache.ignite.lang.IgniteInClosure;
 import org.apache.ignite.metric.MetricRegistry;
 import org.apache.ignite.plugin.extensions.communication.IoPool;
 import org.apache.ignite.spi.systemview.view.StripedExecutorTaskView;
-import org.apache.ignite.thread.IgniteStripedThreadPoolExecutor;
-import org.apache.ignite.thread.IgniteThreadPoolExecutor;
-import org.apache.ignite.thread.SameThreadExecutor;
 import org.jetbrains.annotations.Nullable;
 
 import static org.apache.ignite.IgniteSystemProperties.IGNITE_STARVATION_CHECK_INTERVAL;
