+++ modules/core/src/main/java/org/apache/ignite/internal/thread/package-info.java
@@ -20,4 +20,4 @@
  * Contains threads-related utility classes.
  */
 
-package org.apache.ignite.thread;
+package org.apache.ignite.internal.thread;
