+++ modules/core/src/main/java/org/apache/ignite/internal/thread/pool/OperationContextAwareStripedThreadPoolExecutor.java
@@ -18,7 +18,6 @@
 package org.apache.ignite.internal.thread.pool;
 
 import org.apache.ignite.internal.thread.context.function.OperationContextAwareRunnable;
-import org.apache.ignite.thread.IgniteStripedThreadPoolExecutor;
 
 /** */
 public class OperationContextAwareStripedThreadPoolExecutor extends IgniteStripedThreadPoolExecutor {
