+++ modules/core/src/test/java/org/apache/ignite/marshaller/GridMarshallerResourceBean.java
@@ -32,8 +32,8 @@ import org.apache.ignite.compute.ComputeJobContext;
 import org.apache.ignite.compute.ComputeLoadBalancer;
 import org.apache.ignite.compute.ComputeTaskSession;
 import org.apache.ignite.internal.managers.loadbalancer.GridLoadBalancerAdapter;
+import org.apache.ignite.internal.thread.IgniteThreadFactory;
 import org.apache.ignite.logger.java.JavaLogger;
-import org.apache.ignite.thread.IgniteThreadFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.GenericApplicationContext;
