+++ modules/core/src/test/java/org/apache/ignite/testsuites/IgniteUtilSelfTestSuite.java
@@ -20,6 +20,10 @@ package org.apache.ignite.testsuites;
 import org.apache.ignite.internal.IgniteVersionUtilsSelfTest;
 import org.apache.ignite.internal.management.api.CommandUtilsSelfTest;
 import org.apache.ignite.internal.pagemem.impl.PageIdUtilsSelfTest;
+import org.apache.ignite.internal.thread.GridThreadTest;
+import org.apache.ignite.internal.thread.pool.GridThreadPoolExecutorServiceSelfTest;
+import org.apache.ignite.internal.thread.pool.IgniteThreadPoolSizeTest;
+import org.apache.ignite.internal.thread.pool.ThreadPoolMetricsTest;
 import org.apache.ignite.internal.util.BasicRateLimiterTest;
 import org.apache.ignite.internal.util.DistributedProcessClientAwaitTest;
 import org.apache.ignite.internal.util.DistributedProcessCoordinatorLeftTest;
@@ -52,10 +56,6 @@ import org.apache.ignite.spi.discovery.ClusterMetricsSelfTest;
 import org.apache.ignite.spi.discovery.ClusterMetricsSnapshotSerializeCompatibilityTest;
 import org.apache.ignite.spi.discovery.ClusterMetricsSnapshotSerializeSelfTest;
 import org.apache.ignite.spi.discovery.ClusterRebalancedMetricTest;
-import org.apache.ignite.thread.GridThreadPoolExecutorServiceSelfTest;
-import org.apache.ignite.thread.GridThreadTest;
-import org.apache.ignite.thread.IgniteThreadPoolSizeTest;
-import org.apache.ignite.thread.ThreadPoolMetricsTest;
 import org.apache.ignite.util.GridConcurrentLinkedDequeMultiThreadedTest;
 import org.apache.ignite.util.GridIntListSelfTest;
 import org.apache.ignite.util.GridLogThrottleTest;
